+++ __tests__/ApplicationTest.js
@@ -1,55 +1,56 @@
-const MissionUtils = require("@woowacourse/mission-utils");
 const App = require("../src/App");
 const BridgeMaker = require("../src/BridgeMaker");
+const { mockQuestions, mockRandoms, getLogSpy, getOutput, expectLogContains,
+  expectBridgeOrder,runException,}  = require("..//util/TestUtil");
 
-const mockQuestions = (answers) => {
-  MissionUtils.Console.readLine = jest.fn();
-  answers.reduce((acc, input) => {
-    return acc.mockImplementationOnce((_, callback) => {
-      callback(input);
-    });
-  }, MissionUtils.Console.readLine);
-};
+// const mockQuestions = (answers) => {
+//   MissionUtils.Console.readLine = jest.fn();
+//   answers.reduce((acc, input) => {
+//     return acc.mockImplementationOnce((_, callback) => {
+//       callback(input);
+//     });
+//   }, MissionUtils.Console.readLine);
+// };
 
-const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickNumberInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickNumberInRange);
-};
+// const mockRandoms = (numbers) => {
+//   MissionUtils.Random.pickNumberInRange = jest.fn();
+//   numbers.reduce((acc, number) => {
+//     return acc.mockReturnValueOnce(number);
+//   }, MissionUtils.Random.pickNumberInRange);
+// };
 
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, "print");
-  logSpy.mockClear();
-  return logSpy;
-};
+// const getLogSpy = () => {
+//   const logSpy = jest.spyOn(MissionUtils.Console, "print");
+//   logSpy.mockClear();
+//   return logSpy;
+// };
 
-const getOutput = (logSpy) => {
-  return [...logSpy.mock.calls].join("");
-};
+// const getOutput = (logSpy) => {
+//   return [...logSpy.mock.calls].join("");
+// };
 
-const runException = (inputs) => {
-  mockQuestions(inputs);
-  const logSpy = getLogSpy();
-  const app = new App();
+// const runException = (inputs) => {
+//   mockQuestions(inputs);
+//   const logSpy = getLogSpy();
+//   const app = new App();
 
-  app.play();
+//   app.play();
 
-  expectLogContains(getOutput(logSpy), ["[ERROR]"]);
-};
+//   expectLogContains(getOutput(logSpy), ["[ERROR]"]);
+// };
 
-const expectLogContains = (received, logs) => {
-  logs.forEach((log) => {
-    expect(received).toEqual(expect.stringContaining(log));
-  });
-};
+// const expectLogContains = (received, logs) => {
+//   logs.forEach((log) => {
+//     expect(received).toEqual(expect.stringContaining(log));
+//   });
+// };
 
-const expectBridgeOrder = (received, upside, downside) => {
-  const upsideIndex = received.indexOf(upside);
-  const downsideIndex = received.indexOf(downside);
+// const expectBridgeOrder = (received, upside, downside) => {
+//   const upsideIndex = received.indexOf(upside);
+//   const downsideIndex = received.indexOf(downside);
 
-  expect(upsideIndex).toBeLessThan(downsideIndex);
-};
+//   expect(upsideIndex).toBeLessThan(downsideIndex);
+// };
 
 describe("다리 건너기 테스트", () => {
   test("다리 생성 테스트", () => {
