+++ src/App.js
@@ -15,7 +15,7 @@ class App {
   }
 }
 
-const app = new App();
+const app = new App(); 
 app.play();
 
 module.exports = App;
